requirements.txt
langchain-google-genai
streamlit
python-dotenv
pandas
sqlalchemy